+++ test/Core/Reception/ReceiverStackBuilderTest.php
@@ -27,7 +27,6 @@ class ReceiverStackBuilderTest extends TestCase
     {
         $stackBuilder = new ReceiverStackBuilder(
             new ReceiverCollection(),
-            new ReceiverCollection(),
             FooBarEntity::class,
             $this->createMock(LoggerInterface::class),
         );
@@ -91,8 +90,7 @@ class ReceiverStackBuilderTest extends TestCase
         $receiver2->method('supports')->willReturn(FooBarEntity::class);
 
         $stackBuilder = new ReceiverStackBuilder(
-            new ReceiverCollection([$receiver1, $receiver2]),
-            new ReceiverCollection([$receiver2]),
+            new ReceiverCollection([$receiver1, $receiver2, $receiver2]),
             FooBarEntity::class,
             $this->createMock(LoggerInterface::class),
         );